feat : Commnet 엔드포인트 자원 소유권 확인 로직 추가
- 각각의 엔드포인트에서 자원 소유권 확인 및 필요 시 소유자인지 속성 추가
- 엔드포인트에서 memberPublicId를 AuthenticationPrincipal로 받을 수 있게
설정
- service layer, entity에 소유권 확인 로직 추가
Open #10

+++ src/main/java/com/example/community/service/CommentService.java
@@ -6,15 +6,18 @@ import com.example.community.repository.CommentQueryRepository;
 import com.example.community.repository.MemberQueryRepository;
 import com.example.community.repository.PostQueryRepository;
 import com.example.community.service.dto.CommentCreateRequestDto;
+import com.example.community.service.dto.CommentDeleteRequestDto;
 import com.example.community.service.dto.CommentDetailResponseDto;
 import com.example.community.service.dto.CommentUpdateRequestDto;
 import com.example.community.service.dto.PageCommentRequestDto;
 import com.example.community.service.entity.Comment;
 import com.example.community.service.exception.BoardNotExistException;
 import com.example.community.service.exception.MemberNotFoundException;
+import com.example.community.service.exception.NotResourceOwnerException;
 import com.example.community.service.exception.PostNotExistException;
-import java.time.OffsetDateTime;
+import io.jsonwebtoken.lang.Assert;
 import java.util.UUID;
+import java.util.function.Predicate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
@@ -31,7 +34,32 @@ public class CommentService {
   private final MemberQueryRepository memberQueryRepository;
 
   public Page<CommentDetailResponseDto> findComments(PageCommentRequestDto dto) {
-    return commentQueryRepository.findComments(dto);
+    Page<CommentDetailResponseDto> page = commentQueryRepository.findComments(dto);
+
+    if (isNullValue(dto.getMemberPublicId())) {
+      return page;
+    }
+
+    page.stream()
+        .filter(isOwnerOfComment(dto.getMemberPublicId()))
+        .forEach(this::assignOwnership);
+
+    return page;
+  }
+
+  private boolean isNullValue(UUID memberPublicId) {
+    return memberPublicId.getMostSignificantBits() == 0
+        && memberPublicId.getLeastSignificantBits() == 0;
+  }
+
+  private void assignOwnership(CommentDetailResponseDto dto) {
+    Assert.isTrue(!dto.isOwner(), "소유권은 한 번만 변경 가능합니다.");
+
+    dto.setIsOwner(true);
+  }
+
+  private Predicate<CommentDetailResponseDto> isOwnerOfComment(UUID requestMemberPublicId) {
+    return comment -> comment.getMemberPublicId() == requestMemberPublicId;
   }
 
   @Transactional
@@ -53,18 +81,26 @@ public class CommentService {
     Comment comment = commentJpaRepository.findByBoardPublicIdAndPostPublicIdAndPublicId(
         dto.getBoardPublicId(), dto.getPostPublicId(), dto.getCommentPublicId());
 
+    if (comment.isNotOwner(dto.getMemberPublicId())) {
+      throw NotResourceOwnerException.ofComment();
+    }
+
     comment.changeContent(dto.getContent());
 
     return dto.getCommentPublicId();
   }
 
   @Transactional
-  public UUID deleteComment(UUID boardPublicId, UUID postPublicId, UUID commentPublicId) {
+  public UUID deleteComment(CommentDeleteRequestDto dto) {
     Comment comment = commentJpaRepository.findByBoardPublicIdAndPostPublicIdAndPublicId(
-        boardPublicId, postPublicId, commentPublicId);
+        dto.getBoardPublicId(), dto.getPostPublicId(), dto.getCommentPublicId());
+
+    if (comment.isNotOwner(dto.getMemberPublicId())) {
+      throw NotResourceOwnerException.ofComment();
+    }
 
     commentJpaRepository.remove(comment);
 
-    return commentPublicId;
+    return dto.getCommentPublicId();
   }
 }
